+++ CHANGELOG.md
@@ -35,6 +35,7 @@
 - Can't find `performanceTesting.jar` when building against Android Studio 242+. JetBrains/intellij-platform-gradle-plugin#1738
 - Custom `runIde` task do not find the right runtime, if `useInstaller` is `false`. JetBrains/intellij-platform-gradle-plugin#1827
 - PluginVerifier doesn't honor gradle offline mode. JetBrains/intellij-platform-gradle-plugin#1820
+- Make  `intellij.rider` bundled module automatically available for Rider. JetBrains/intellij-platform-gradle-plugin#1774
 
 ## [2.1.0]